新增 unlockConstraint 布局解除约束功能
- 在 `RowListLayout` 和 `ColumnListLayout` 中增加解约束逻辑。
- 优化 `CollectionConfigWrapper` 的悬浮文本生成逻辑，提取为可配置的 `hoverEntryToString`。
- 更新 `showFirstConfigInContainer` 默认值为 `false`。

+++ src/client/kotlin/moe/forpleuvoir/ibukigourd/gui/base/layout/LinearLayout.kt
@@ -74,7 +74,7 @@ interface RowLayout : LinearLayout<Arrangement.Vertical, Alignment.Horizontal> {
         //使用的高度
         var usedHeight = arrangement.spacing * measurables.lastIndex
         //总权重
-        val totalWidget = parentDatas.sumOf { it.weight.toInt() }
+        val totalWidget = parentDatas.sumOf { it.weight }
 
         //优先计算固定最小尺寸的组件的占用
         var minimumOccupiedHeight = 0f
